// src/lib-tiers/models/Role.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Permission } from './Permission';
export type Role = {
    id?: string;
    name?: string;
    description?: string;
    permissions?: Array<Permission>;
};